rename ChatsScreen to HomeScreen and related components
