--- core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/navigation/NavigationConstants.kt
@@ -1,5 +1,0 @@
-package net.thechance.mena.core_chat.presentation.navigation
-
-object NavigationConstants {
-    const val IS_SYNC_SUCCESS = "is_sync_success"
-}
