+++ core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/navigation/Routes.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 interface ChatRoute
 
 @Serializable
-data object ChatsRoute : ChatRoute
+data object HomeRoute : ChatRoute
 
 @Serializable
 data object WalletRoute : ChatRoute
