+++ core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/screen/chats/HomeScreenState.kt
@@ -1,6 +1,6 @@
 package net.thechance.mena.core_chat.presentation.screen.chats
 
-data class ChatsScreenState(
+data class HomeScreenState(
     val isLoading: Boolean = false,
     val isSynced: Boolean = false
 )
