+++ core_chat_presentation/src/commonMain/kotlin/net/thechance/mena/core_chat/presentation/screen/chats/HomeViewModel.kt
@@ -7,10 +7,10 @@ import net.thechance.mena.core_chat.presentation.navigation.SyncContactsRoute
 import net.thechance.mena.core_chat.presentation.navigation.WalletRoute
 import net.thechance.mena.core_chat.presentation.shared.BaseViewModel
 
-class ChatsViewModel(
+class HomeViewModel(
     private val contactsRepository: ContactsRepository,
     effector: ChatEffector
-) : BaseViewModel<ChatsScreenState>(ChatsScreenState(), effector) {
+) : BaseViewModel<HomeScreenState>(HomeScreenState(), effector) {
 
     fun onNewChatClicked() {
         tryToExecute(
