+++ core_chat_presentation/src/commonTest/kotlin/net/thechance/mena/core_chat/presentation/screen/contacts/ContactsViewModelTest.kt
@@ -34,7 +34,7 @@ import net.thechance.mena.core_chat.domain.model.PagedData
 import net.thechance.mena.core_chat.domain.repository.ChatRepository
 import net.thechance.mena.core_chat.domain.repository.ContactsRepository
 import net.thechance.mena.core_chat.presentation.navigation.ChatEffector
-import net.thechance.mena.core_chat.presentation.navigation.NavigationConstants.IS_SYNC_SUCCESS
+import net.thechance.mena.core_chat.presentation.screen.syncContacts.IS_SYNC_SUCCESS
 import net.thechance.mena.core_chat.presentation.utils.UiText
 import kotlin.test.AfterTest
 import kotlin.test.BeforeTest
